\name{cms-package}
\alias{cms-package}
\alias{cms}
\docType{package}
\title{
\packageTitle{cms}
}
\description{
\packageDescription{cms}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cms}
\packageIndices{cms}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cms}

Maintainer: \packageMaintainer{cms}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
