% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_clusters.R
\name{cms_clusters}
\alias{cms_clusters}
\title{cms algorithm}
\usage{
cms_clusters(raw = raw, runs = 100, emptysize = 0.2,
  trainsize = 0.8, idvariable = "animal_id", varstart = 14,
  exclude = "Seizures_n", scorevars = NA)
}
\arguments{
\item{file}{path and filename as a string (e.g. C:/testfolder/testdata.txt)}

\item{remove_this}{specification of columheaders that shall be removed from the data}
}
\description{
The algorithm takes the cleaned input data, does some curation for missing values and then calculates the composite measures schemes.
The function can be used in two ways: 1. with the scorevars set to NA the data are screened for the most prominent feature contributions; 2. 
the so found feature sets can be specified, e.g., scorevars=c("Sacc_pref", "social_interaction", "burrowing_rat", "openfield_rat"). This will
result in the calculation of a severity-based cluster distribution analysis using the input variables.
}
