% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cms.R
\name{plot_cms}
\alias{plot_cms}
\title{Plot cms Clusters}
\usage{
plot_cms(cmsresult = NULL, rotateX = NULL)
}
\arguments{
\item{cmsresult}{requires the output object from the cms function}

\item{rotateX}{rotates the x labels (default=NULL; 45 means 45 degrees)}
}
\value{
Data frame with the cluster distributions/severity attributions.+
}
\description{
The function is a wrapper for subgroup filtering of the cms output and 
lotting the cluster distributions.
}
