% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\docType{package}
\name{cms}
\alias{cms}
\title{cms: Tools for Calculating CMS Medicare Reimbursement}
\description{
Uses the CMS API to provide useRs databases containing yearly
Medicare reimbursement rates in the United States. Data can be acquired
for the entire United States or only for specific localities. Currently,
support is only provided for the Medicare Physician Fee Schedule (MPFS),
but support will be expanded for other CMS databases in future versions.
}
