% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_analysis.R
\name{cms_analysis}
\alias{cms_analysis}
\title{Takes in the results from the cms_clusters function (frequency distributions of each run of the 80% subsample of the training data) and calculates
general attributes as well as graphical distribution outputs.}
\usage{
cms_analysis(raw, thresholds, cluster_distribution, savepath = NA)
}
\arguments{
\item{raw}{requires the raw data list (also input in the cms_load function)}

\item{thresholds}{specification of the threshold values from the listed output of cms_clusters (e.g. cms_cl[[1]] and cms_cl[[2]])}

\item{cluster_distribution}{output from the cms_cluster function (IMPORTANT: scorevars requires a vector of input variables!)}

\item{savepath}{include the path for the baplot here (no name spec for the plot necessary, it is called "Cluster_distros.png")}
}
\value{
data.frame
}
\description{
Takes in the results from the cms_clusters function (frequency distributions of each run of the 80% subsample of the training data) and calculates
general attributes as well as graphical distribution outputs.
}
