% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_clusters.R
\name{cms_clusters}
\alias{cms_clusters}
\title{Cluster and frequency distribution of composite measures for severity attribution}
\usage{
cms_clusters(raw = raw, runs = 100, emptysize = 0.2,
  trainsize = 0.8, idvariable = "animal_id", varstart = 14,
  exclude = "Seizures_n", scorevars = NA)
}
\arguments{
\item{raw}{raw data (unprocessed or from cms_load)}

\item{runs}{number of random samplings for the PCA-based feature determination}

\item{emptysize}{fraction of empty data that will be removed}

\item{trainsize}{size of the "training data" in the fold subsamplings}

\item{idvariable}{column header of the animal id variable in the data}

\item{varstart}{column number with the first variable to start with - there shall be no general infomation on the right side of this number}

\item{exclude}{naming column headers that shall be excluded from the data set}

\item{scorevars}{if NA (default) the function will determine the most prominent variables in the sample data; NA can also be replaced by a vector of variables}
}
\value{
A list with composite score thresholds and k-means cluster attributions.
}
\description{
The algorithm takes the cleaned input data, does some curation for missing values and then calculates the composite measures schemes.
The function can be used in two ways: 1. with the scorevars set to NA the data are screened for the most prominent feature contributions; 2. 
the so found feature sets can be specified, e.g., scorevars=c("Sacc_pref", "social_interaction", "burrowing_rat", "openfield_rat"). This will
result in the calculation of a severity-based cluster distribution analysis using the input variables.
}
