% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episet_full.R
\docType{data}
\name{episet_full}
\alias{episet_full}
\title{Full data set rat epilepsy experiment}
\format{A 943 x 72 data frame with rat epilepsy data.}
\usage{
data(episet_full)
}
\description{
This cool data set includes it all...
}
\examples{
data(episet_full)
head(episet_full)
}
\keyword{datasets}
